extract wishlist functions into module
